requirements.txt
Mastodon.py
python-dotenv
transformers
torch         # Needed for many Hugging Face models
